+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes of the Respect\Validation releases are documented in this fi
 - Create "KeyValue" rule (#441)
 - Create "Mimetype" rule (#361)
 - Create "NotBlank" rule (#443)
+- Create "NotOptional" rule (#448)
 - Create "Optional" rule (#423)
 - Create "ResourceType" rule (#397)
 - Create "ScalarVal" rule (#397)